==> R/01_load.qmd <==
## Source functions
```{r}
source("../R/99_proj_func.R")
```

## Loading packages
```{r}
list.of.packages <- c("tidyverse", 
                      "Racmacs", 
                      "viridis", 
                      "ggrepel", 
                      "svglite", 
                      "knitr",
                      "rstudioapi", 
                      "kableExtra",
                      "tinytex",
                      "dbscan", #Future perspectives
                      "ggforce",
                      "RColorBrewer",
                      "quarto"
                      )
new.packages <- list.of.packages[!(
  list.of.packages %in% installed.packages()[
    ,"Package"
    ]
  )]
if(length(new.packages)) 
  install.packages(new.packages)
# Special handling for tinytex if not installed
if (!tinytex::is_tinytex()) {
  tinytex::install_tinytex()
}
############################################################################################################
# We are aware that this code may not install all packages, but only the first one if multiple are abscent.#
############################################################################################################
```
## Libraries

```{r}
#| eval: true
#| echo: false
#| message: false
#| warning: false

library("tidyverse")
library("readxl")
library("rstudioapi")
```

# Antigenic Cartography - COVID

Antigen names; H69775 (Wuhan), Alpha (B.1.1.7), Beta (B.1.351), Gamma (P.1), Delta B.1.617.2, Delta AY.33, Delta AY.4, Delta AY.4.2, Delta B.1.617.2 and E484K Omicron

## Generate data directories
```{r}
make_dir("../data")
make_dir("../data/_raw")
```

## Loading excel file

```{r}
# File chooser
file_path <- rstudioapi::selectFile(caption = "Select an Excel file")

# Read excel file
excel_data <- readxl::read_excel(file_path)

  # Check succesful load
if (!is.null(excel_data)) {
  message("Data loaded successfully.")
} else {
  message("Failed to load data.")
}

# If named "NA", then rename to neuraminidase
excel_data <- excel_data |>
  mutate(across(6, ~ifelse(row_number() == 4 & .x == "NA", "neuraminidase", .x))) |>
  mutate(across(where(is.character), ~na_if(.x, "NA")))

## Write csv excel_data file
write_csv(excel_data, "../data/excel_csv.csv.gz")
```

## Loading data
```{r}
#| echo: false
#| message: false
#| output: false

## Load csv titer_table file
titer_table <- read_csv(file = "../data/excel_csv.csv.gz")
```

## Writing files
```{r}
## Write tsv titer_table file
write_tsv(titer_table,
            file = "../data/titer_dirty.tsv.gz")
```
