==> R/03_augment.qmd <==
---
title: "Augment data"
format: html
editor: source
author: authors
---

### Source functions

```{r}
source("../R/99_proj_func.R")
```

### Library

```{r}
#| eval: true
#| echo: true

library("tidyverse")
library("Racmacs")
```

### Loading

```{r}
#| warning: false

ag_names <- read_tsv(file = "../data/ag_names.tsv")
titer_clean <- read_tsv(file = "../data/titer_clean.tsv")
titer_clean <- as.data.frame(titer_clean)
ag_names <- ag_names[["value"]]
row.names(titer_clean) <- ag_names
```

### Augment

```{r}
############################
#   Creating map object    #
############################
set.seed(696969)

options(RacOptimizer.num_cores = parallel::detectCores()) # CPU cores available

map <- acmap(
  titer_table = titer_clean
)

# Map optimizations
map <- optimizeMap(
  map,
  number_of_dimensions    = 2,        # 2D map
  number_of_optimizations = 1000,
  minimum_column_basis    = "none",
  options = list(ignore_disconnected = TRUE))
```

### map file

```{r}
# Define 'results' directory
results_directory <- file.path(getwd(), "../data")

custom_filename <- "map.ace"  
file_path <- file.path("../data", custom_filename)

# Sace map
save.acmap(
  map,
  file_path)
```
