==> R/04_analysis.qmd <==
---
title: "Analyse data"
format: html
editor: source
author: authors
---

### Source functions
```{r}
source("../R/99_proj_func.R")
```

### Library
```{r}
#| eval: true
#| echo: true
#| warning: false

library("tidyverse")
library("Racmacs")
library("viridis")
library("ggrepel")
library("svglite")


```

### Loading
```{r}
file_path <- file.path(getwd(), "../data/map.ace")

map <- read.acmap(
  file_path,
  optimization_number = NULL,
  sort_optimizations = FALSE,
  align_optimizations = FALSE
)

# Attribute tables
attr_table_ag <- read_tsv(file = "../data/attr_table_ag.tsv")
attr_table_sr <- read_tsv(file = "../data/attr_table_sr.tsv")
table_info <- read_tsv(file = "../data/table_info.tsv")
```

## Extracting information and creating colors
```{r}
ag_coords <- agCoords(map)  # Antigen coordinates
sr_coords <- srCoords(map)  # Serum coordinates

############################
# Combine ag and sr data   #
############################
ag_df <- as.data.frame(ag_coords) |>
  mutate(antigen = rownames(ag_coords),
         type = 'AG',
         antigenic_group = attr_table_ag[["Antigenic group"]],
         serum_exposure_homolog = attr_table_ag[["Serum exposure homolog"]])

### Clustering (Futute Perspectives)
clustering_result <- dbscan::dbscan(
  ag_df[, c("V1", "V2")],
  eps = 0.7, # Max distance between two samples for one to be considered part of cluster
  minPts = 2) # Number of samples in a neighborhood for a point to be considered a core point

ag_df <- ag_df |> 
  mutate(cluster = clustering_result$cluster)

# Combine ag and sr continued
sr_df <- as.data.frame(sr_coords) |>
  mutate(serum = rownames(sr_coords),
         type = 'SR')

dd <- bind_rows(ag_df, 
                sr_df) 

dd <- dd |>
  mutate(
    serum_group = attr_table_sr$"Serum exposure antigen"[match(dd$serum, 
                                                  attr_table_sr$"sera_name_with_DPI")]
    )
dd <- dd[is.finite(dd$V1) & is.finite(dd$V2), ]

############################
#     Unique colouring     #
############################
unique_AG <- dd |> 
  filter(type == "AG") |> 
  pull(antigen) |> 
  unique()

colors_for_AG <- setNames(
  viridis::viridis(
    length(unique_AG), 
    option = "turbo"), # color-blind friendly palette
  unique_AG)

colors_for_AG <- sapply(
  colors_for_AG, 
  adjustcolor, 
  alpha.f = 0.8) # Add slight transparency

dd <- dd |> mutate(color = colors_for_AG[antigen])

# Define antigenic groups
unique_antigenic_groups <- unique(dd$antigenic_group)

new_groups <- unique_antigenic_groups[
  !unique_antigenic_groups 
  %in% 
    names(colors_for_AG)]

# Color antigenic groups
if(length(new_groups) > 0) {
  new_colors <- viridis::viridis(
    length(new_groups),
    option = "turbo")
  
  colors_for_AG <- c(colors_for_AG, 
                     setNames(new_colors, 
                              new_groups))
}

# Color serum accordingly to homolog ag
unique_serum_exposures <- unique(attr_table_sr$"Serum exposure antigen")
# ag_df and antigenic_group to color
homolog_colors <- dd |> 
  filter(type == "AG") |> 
  select(serum_exposure_homolog, 
         color) |>
  distinct() |>
  # Ensure there's a row for each antigenic group mentioned in attr_table_sr
  filter(serum_exposure_homolog 
         %in% 
           unique_serum_exposures)

colors_for_SR <- setNames(homolog_colors$color, 
                          homolog_colors$serum_exposure_homolog)

#### Future Perspective ###
library("dbscan")       # (Futute Perspectives)
library("ggforce")      # (Futute Perspectives)
library("RColorBrewer") # (Futute Perspectives)
### Clustering (Futute Perspectives)
num_clusters <- max(ag_df$cluster[ag_df$cluster > 0])
if(num_clusters <= 8) {
  cluster_colors <- RColorBrewer::brewer.pal(num_clusters, "Dark2")
} else {
  # More than 8 clusters
  cluster_colors <- rep(RColorBrewer::brewer.pal(8, "Dark2"), # repeating palette
                        length.out = num_clusters)
}
if (!is.null(num_clusters) && length(num_clusters) > 0 && num_clusters > 0) {
  names(cluster_colors) <- as.character(seq_len(num_clusters))
}

combined_colors <- c(cluster_colors, # Clustering (Futute Perspectives)
                     colors_for_AG, 
                     colors_for_SR, 
                     "Other" = "grey")
```

## base_map
```{r}
############################
# Generating Antigenic Map #
############################
base_map <- ggplot(dd) +
  geom_point(data = filter(dd, type == "SR", !is.na(serum_group)), 
             aes(x = V1, y = V2, shape = type, color = serum_group), 
             size = 2, shape = 0, stroke = 0.4, alpha = 0.7) +  
  geom_point(data = filter(dd, is.na(serum_group), type == "SR"),  
             aes(x = V1, y = V2, color = "Other"),  
             shape = 0, size = 2, alpha = 0.7) +
  scale_y_reverse(breaks = seq(floor(range(dd$V2))[1], 
                               ceiling(range(dd$V2))[2]),
                  expand = expansion(mult = 0, add = 1.0)) +
  scale_x_continuous(breaks = seq(floor(range(dd$V1))[1], 
                                  ceiling(range(dd$V1))[2]),
                     expand = expansion(mult = 0, add = 1.0)) +

  coord_fixed() + 
  scale_size_identity() + 
  
  scale_fill_manual(values = combined_colors) +
  scale_color_manual(values = combined_colors) +
  
  theme_minimal() +
  labs(title = paste0("Antigenic map of ", table_info |> pull(virus_type), " variants"),
       subtitle = paste0("Inhibition titers for ", table_info |> pull(viral_protein), 
                         " analyzed via ", table_info |> pull(assay_type)),
       x = "", y = "", 
       color = "Serum group") +
  theme(axis.text.x = element_blank(), 
        axis.text.y = element_blank(),
        plot.background = element_rect(fill = "white"), 
        panel.background = element_rect(fill = "white"), 
        panel.grid.major = element_line(), 
        panel.grid.minor = element_blank(), 
        legend.title = element_text(margin = margin(l = 1, unit = "pt")),
        legend.text = element_text(size = 8),
        panel.border = element_rect(colour = "azure4", 
                                    fill = NA, 
                                    size = 1.5), 
        plot.margin = margin(t = 20, 
                             r = 20, 
                             b = 10, 
                             l = 10, 
                             unit = "pt"),
        plot.subtitle = element_text(margin = margin(t = 2, 
                                                     b = 10, 
                                                     l = 10,
                                                     r = 5,
                                                     unit = "pt"),
                                     size = 10),
        plot.title = element_text(face = "bold",
                                  margin = margin(l = 10, 
                                                    unit = "pt")), 
        plot.title.position = "plot",
        legend.margin= margin(l = 10, unit = "pt"),
        legend.position = "right")
```

## Legend plot
```{r}
##########
# Legend #
##########
ag_map_1 <- base_map +
  geom_point(data = filter(dd, type == "AG"), 
             aes(x = V1, y = V2, fill = antigen), 
             shape = 21, size = 6, color = "black") +
  labs(fill = "Antigen") + 
  guides(fill = guide_legend(override.aes = list(size = 3)), 
         shape = "none", size = "none",
         color = guide_legend(override.aes = list(size = 2, stroke = 0.8)))
```

## Label plot
```{r}
##########
# Labels #
##########
ag_map_2 <- base_map +
  geom_point(data = filter(dd, type == "AG"), 
             aes(x = V1, y = V2, fill = antigen), 
             shape = 21, size = 6, color = "black") +
  
  geom_text_repel(aes(x = V1, 
                      y = V2, 
                      label = antigen),
                  size = 3, 
                  box.padding = unit(0.4, 
                                     "lines"), 
                  point.padding = unit(1.05, "lines"),
                  segment.linetype = 1, 
                  min.segment.length = 0.4, 
                  arrow = arrow(length = unit(0.015, "npc")),
                  segment.curvature = 0, 
                  segment.alpha = 0.7, 
                  force_pull = 1.0, 
                  force = 1.0, 
                  max.iter = 10000,
                  xlim = c(-NA, NA), 
                  ylim = c(-NA, NA), 
                  bg.color = "white", 
                  bg.r = 0.10) +
  labs(fill = "Antigen") + 
  guides(fill = FALSE, 
         color = guide_legend(override.aes = list(size = 2, stroke = 0.8)))
```

## Plain plot
```{r}
##########
# Plain  #
##########
ag_map_3 <- base_map +
  geom_point(data = filter(dd, type == "AG"), 
             aes(x = V1, y = V2, fill = antigen), 
             shape = 21, size = 6, color = "black") +
  labs(fill = "Antigen") + 
  guides(fill = guide_legend(override.aes = list(size = 3)), shape = "none", size = "none",
         color = guide_legend(override.aes = list(size = 2, stroke = 0.8))) +
  theme(legend.position = "none")
```

```{r}
##########
# Grouped#
##########
ag_map_4 <- base_map + 
  geom_point(data = filter(dd, type == "AG"), 
             aes(x = V1, y = V2, fill = antigenic_group), 
             shape = 21, size = 6, color = "black") +
  labs(fill = "Antigenic group") + 
  guides(fill = guide_legend(override.aes = list(size = 3)), shape = "none", size = "none",
         color = guide_legend(override.aes = list(size = 2, stroke = 0.8)))

ag_counts <- dd |>
  filter(type == "AG") |>
  count(antigenic_group)

########################
# Grouped with Labels #
#######################
# Define antigenic groups
ag_groups_with_multiple_entries <- ag_counts |>
  filter(n > 1) |>
  pull(antigenic_group)
# Add to dd
dd_multiple_entries <- dd |>
  filter(antigenic_group 
         %in% 
           ag_groups_with_multiple_entries)

### ___________________________________________
### Clustering         (Futute Perspectives)
dd_cluster <- dd |> 
  filter(type == "AG") |> 
  filter(cluster != 0) 
### ___________________________________________

ag_map_5 <- base_map +
  geom_point(data = filter(dd, 
                           type == "AG"), 
             aes(x = V1, y = V2, 
                 fill = antigenic_group), 
             shape = 21, 
             size = 6, 
             color = "black") +
  
  geom_text_repel(data = dd_multiple_entries,
                  aes(x = V1, 
                      y = V2, 
                      label = antigen),
                  size = 3, 
                  box.padding = unit(0.35, "lines"), 
                  point.padding = unit(0.8, "lines"),
                  segment.linetype = 1, 
                  min.segment.length = 0.4, 
                  arrow = arrow(length = unit(0.015, "npc")),
                  segment.curvature = 0, 
                  segment.alpha = 0.7, 
                  force_pull = 1.0, 
                  force = 1.0, 
                  max.iter = 10000,
                  xlim = c(-NA, NA), 
                  ylim = c(-NA, NA), 
                  bg.color = "white", 
                  bg.r = 0.10) +
  labs(fill = "Antigenic group") + 
  
  guides(fill = guide_legend(override.aes = list(size = 3)), 
         color = guide_legend(override.aes = list(size = 2, stroke = 0.8)))

```


Saving the relevant plots

```{r}
############################
# Save the map (.svg.tiff) #
############################
make_dir("../results")

base_filename <- paste0("ac_map_", 
                        table_info |>
                          pull(virus_type), 
                        "_", Sys.Date())

base_folder_path <- "../results"
folder_name <- paste0("ac_map_", table_info |> pull(virus_type), "_", Sys.Date())
final_folder_path <- create_unique_folder(base_folder_path, folder_name)


## JPG
ggsave(
  plot = ag_map_1,
  filename = paste0(base_filename, ".jpg"),
  path = final_folder_path,
  device = "jpg",
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_2,
  filename = paste0(base_filename,"_label", ".jpg"),
  path = final_folder_path,
  device = "jpg",
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_3,
  filename = paste0(base_filename,"_plain", ".jpg"),
  path = final_folder_path,
  device = "jpg",
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_4,
  filename = paste0(base_filename,"_grouped", ".jpg"),
  path = final_folder_path,
  device = "jpg",
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_5,
  filename = paste0(base_filename,"_grouped_label", ".jpg"),
  path = final_folder_path,
  device = "jpg",
  width = 20,
  height = 12,
  units = "cm",
)
## SVG
ggsave(
  plot = ag_map_1,
  filename = paste0(base_filename, ".svg"),
  path = final_folder_path,
  device = "svg",
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_2,
  filename = paste0(base_filename,"_label", ".svg"),
  path = final_folder_path,
  device = "svg",
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_3,
  filename = paste0(base_filename,"_plain", ".svg"),
  path = final_folder_path,
  device = "svg",
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_4,
  filename = paste0(base_filename,"_grouped", ".svg"),
  path = final_folder_path,
  device = "svg",
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_5,
  filename = paste0(base_filename,"_grouped_label", ".svg"),
  path = final_folder_path,
  device = "svg",
  width = 20,
  height = 12,
  units = "cm",
)

## TIFF
ggsave(
  plot = ag_map_1,
  filename = paste0(base_filename, ".tiff"),
  path = final_folder_path,
  device = "tiff",
  dpi = 300, # Set the desired resolution
  width = 20,
  height = 12,
  units = "cm",
)

ggsave(
  plot = ag_map_2,
  filename = paste0(base_filename,"_label", ".tiff"),
  path = final_folder_path,
  device = "tiff",
  dpi = 300, # Set the desired resolution
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_3,
  filename = paste0(base_filename,"_plain", ".tiff"),
  path = final_folder_path,
  device = "tiff",
  dpi = 300, # Set the desired resolution
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_4,
  filename = paste0(base_filename,"_grouped", ".tiff"),
  path = final_folder_path,
  device = "tiff",
  dpi = 300, # Set the desired resolution
  width = 20,
  height = 12,
  units = "cm",
)
ggsave(
  plot = ag_map_5,
  filename = paste0(base_filename,"_grouped_label", ".tiff"),
  path = final_folder_path,
  device = "tiff",
  dpi = 300, # Set the desired resolution
  width = 20,
  height = 12,
  units = "cm",
)

write_tsv(dd,
            file = "../data/dd.tsv")

```
